// src/app/models/auth-model.ts
export interface IForgetModel {
  email: string;
}

// Generated by https://quicktype.io

export interface IForgetPasswordModel {
  password: string;
  userId:   string;
  token:    string;
}
// Generated by https://quicktype.io

export interface IUpdatePassword {
  currentPassword: string;
  newPassword:     string;
}